Movie reader (#7)
Adds a bare-bones Gaffer implementation consisting of a simple MovieReader node. Internally, an AvReader node is used to do the heavy lifting of generating frames from a movie file (although it is mocked currently).

A demuxer has been sketched out and a first implementation proposed, along with some simple tests. The seek functionality remains to be implemented.